move sniffer to its own file
